Test TrustSchema + Term.weight type update

+++ attestation-transformer/src/term.rs
@@ -35,17 +35,17 @@ impl Into<Form> for TermForm {
 	}
 }
 
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct Term {
 	from: String,
 	to: String,
-	weight: u32,
+	weight: f32,
 	domain: u32,
 	form: TermForm,
 }
 
 impl Term {
-	pub fn new(from: String, to: String, weight: u32, domain: u32, is_trust: bool) -> Term {
+	pub fn new(from: String, to: String, weight: f32, domain: u32, is_trust: bool) -> Term {
 		Term {
 			from,
 			to,
@@ -93,7 +93,7 @@ impl Term {
 
 				let from = hex::encode(from_bytes);
 				let to = hex::encode(to_bytes);
-				let weight = u32::from_be_bytes(weight_bytes);
+				let weight = f32::from_be_bytes(weight_bytes);
 				let domain = u32::from_be_bytes(domain_bytes);
 				let form = TermForm::from(form_byte);
 				Term { from, to, weight, domain, form }
@@ -117,7 +117,7 @@ impl Term {
 				let from =
 					String::from_utf8(from_bytes).map_err(|_| AttTrError::SerialisationError)?;
 				let to = String::from_utf8(to_bytes).map_err(|_| AttTrError::SerialisationError)?;
-				let weight = u32::from_be_bytes(weight_bytes);
+				let weight = f32::from_be_bytes(weight_bytes);
 				let domain = u32::from_be_bytes(domain_bytes);
 				let form = TermForm::from(form_byte);
 				Term { from, to, weight, domain, form }
@@ -151,7 +151,7 @@ mod test {
 		let term = Term {
 			from: "did:eth:pkh:90f8bf6a479f320ead074411a4b0e7944ea8c9c1".to_owned(),
 			to: "did:eth:pkh:90f8bf6a479f320ead074411a4b0e7944ea8c9c2".to_owned(),
-			weight: 50,
+			weight: 50.,
 			domain: 67834578,
 			form: TermForm::Trust,
 		};
